Updated output format
