+++ crates/ruff/resources/test/fixtures/pygrep-hooks/PGH001_0.py
@@ -1,6 +1,6 @@
 from ast import literal_eval
 
-eval("3 + 4")  # noqa: S001
+eval("3 + 4")
 
 literal_eval({1: 2})
