+++ crates/ruff/src/rules/flake8_bandit/mod.rs
@@ -12,11 +12,11 @@ mod tests {
     use test_case::test_case;
 
     use crate::registry::Rule;
-    use crate::rules::flake8_bandit::helpers::Severity;
+    use crate::rules::flake8_bandit::settings::Severity;
     use crate::settings::Settings;
     use crate::test::test_path;
 
-    #[test_case(Rule::DenylistCall, Path::new("S001.py"); "S001")]
+    #[test_case(Rule::DeniedFunctionCall, Path::new("S001.py"); "S001")]
     #[test_case(Rule::Assert, Path::new("S101.py"); "S101")]
     #[test_case(Rule::ExecBuiltin, Path::new("S102.py"); "S102")]
     #[test_case(Rule::BadFilePermissions, Path::new("S103.py"); "S103")]
